refactor: 오답 생성 수정 코드 리뷰 반영
