+++ apps/web/src/app/client-shell.tsx
@@ -6,7 +6,7 @@ import { AuthLogoutListener } from "@/shared/apis/auth/auth-logout-listener";
 import QueryProvider from "@/shared/utils/query-provider";
 import AppBarGate from "@/shared/components/app-bar/app-bar-gate";
 import BottomNav from "@/shared/components/bottom-nav/bottom-nav";
-import FabButton from "@/shared/components/button/fab-button/fab-button";
+import CreateWrongFab from "@/app/create-wrong-fab";
 import Splash from "@/shared/components/splash/splash";
 import QueryLoadingOverlay from "@/shared/components/loading/query-loading-overlay";
 
@@ -25,7 +25,7 @@ const ClientShell = ({ children }: ClientShellProps) => {
       <AppBarGate />
       <div key={contentKey}>{children}</div>
       <BottomNav />
-      <FabButton />
+      <CreateWrongFab />
     </QueryProvider>
   );
 };
