+++ apps/web/src/shared/components/button/fab-button/fab-button.tsx
@@ -2,8 +2,6 @@
 
 import type { ButtonHTMLAttributes, MouseEvent } from "react";
 import clsx from "clsx";
-import { usePathname, useRouter, useSearchParams } from "next/navigation";
-import { ROUTES } from "@/shared/constants/routes";
 import Icon from "@/shared/components/icon/icon";
 import type { IconProps } from "@/shared/components/icon/icon";
 import * as styles from "@/shared/components/button/fab-button/fab-button.css";
@@ -12,6 +10,7 @@ type Props = Omit<ButtonHTMLAttributes<HTMLButtonElement>, "children"> & {
   icon?: IconProps["name"];
   iconSize?: number;
   label?: string;
+  onFabClick?: (e: MouseEvent<HTMLButtonElement>) => void;
 };
 
 const FabButton = ({
@@ -20,35 +19,11 @@ const FabButton = ({
   label = "오답 등록하기",
   className,
   type = "button",
-  onClick,
+  onFabClick,
   ...rest
 }: Props) => {
-  const pathname = usePathname();
-  const sp = useSearchParams();
-  const router = useRouter();
-
-  const SHOW_ON_PATHS = [
-    ROUTES.HOME,
-    ROUTES.WRONG.ROOT,
-    ROUTES.GRAPH.ROOT,
-  ] as const;
-
-  if (!SHOW_ON_PATHS.includes(pathname as (typeof SHOW_ON_PATHS)[number]))
-    return null;
-
   const handleClick = (e: MouseEvent<HTMLButtonElement>) => {
-    onClick?.(e);
-    if (e.defaultPrevented) return;
-
-    const qs = sp.toString();
-    const from = `${pathname}${qs ? `?${qs}` : ""}`;
-
-    const nextParams = new URLSearchParams();
-    nextParams.set("step", "1");
-    nextParams.set("total", "4");
-    nextParams.set("from", from);
-
-    router.push(`${ROUTES.WRONG.CREATE}?${nextParams.toString()}`);
+    onFabClick?.(e);
   };
 
   const ariaLabel = rest["aria-label"] ?? label;
